% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_associate_device_with_room}
\alias{alexaforbusiness_associate_device_with_room}
\title{Associates a device with a given room}
\usage{
alexaforbusiness_associate_device_with_room(DeviceArn, RoomArn)
}
\arguments{
\item{DeviceArn}{The ARN of the device to associate to a room. Required.}

\item{RoomArn}{The ARN of the room with which to associate the device. Required.}
}
\description{
Associates a device with a given room. This applies all the settings
from the room profile to the device, and all the skills in any skill
groups added to that room. This operation requires the device to be
online, or else a manual sync is required.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$associate_device_with_room(
  DeviceArn = "string",
  RoomArn = "string"
)
}
}

\keyword{internal}
