% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_contact}
\alias{alexaforbusiness_get_contact}
\title{Gets the contact details by the contact ARN}
\usage{
alexaforbusiness_get_contact(ContactArn)
}
\arguments{
\item{ContactArn}{[required] The ARN of the contact for which to request details.}
}
\description{
Gets the contact details by the contact ARN.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$get_contact(
  ContactArn = "string"
)
}
}

\keyword{internal}
