% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_put_room_skill_parameter}
\alias{alexaforbusiness_put_room_skill_parameter}
\title{Updates room skill parameter details by room, skill, and parameter key
ID}
\usage{
alexaforbusiness_put_room_skill_parameter(RoomArn, SkillId,
  RoomSkillParameter)
}
\arguments{
\item{RoomArn}{The ARN of the room associated with the room skill parameter. Required.}

\item{SkillId}{[required] The ARN of the skill associated with the room skill parameter. Required.}

\item{RoomSkillParameter}{[required] The updated room skill parameter. Required.}
}
\description{
Updates room skill parameter details by room, skill, and parameter key
ID. Not all skills have a room skill parameter.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$put_room_skill_parameter(
  RoomArn = "string",
  SkillId = "string",
  RoomSkillParameter = list(
    ParameterKey = "string",
    ParameterValue = "string"
  )
)
}
}

\keyword{internal}
