% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_start_device_sync}
\alias{alexaforbusiness_start_device_sync}
\title{Resets a device and its account to the known default settings, by
clearing all information and settings set by previous users}
\usage{
alexaforbusiness_start_device_sync(RoomArn, DeviceArn, Features)
}
\arguments{
\item{RoomArn}{The ARN of the room with which the device to sync is associated.
Required.}

\item{DeviceArn}{The ARN of the device to sync. Required.}

\item{Features}{[required] Request structure to start the device sync. Required.}
}
\description{
Resets a device and its account to the known default settings, by
clearing all information and settings set by previous users.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$start_device_sync(
  RoomArn = "string",
  DeviceArn = "string",
  Features = list(
    "BLUETOOTH"|"VOLUME"|"NOTIFICATIONS"|"LISTS"|"SKILLS"|"ALL"
  )
)
}
}

\keyword{internal}
