% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_tag_resource}
\alias{alexaforbusiness_tag_resource}
\title{Adds metadata tags to a specified resource}
\usage{
alexaforbusiness_tag_resource(Arn, Tags)
}
\arguments{
\item{Arn}{[required] The ARN of the resource to which to add metadata tags. Required.}

\item{Tags}{[required] The tags to be added to the specified resource. Do not provide system
tags. Required.}
}
\description{
Adds metadata tags to a specified resource.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$tag_resource(
  Arn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
